separate asdf and stnode
