+++ src/roman_datamodels/_asdf/integration.py
@@ -8,6 +8,6 @@ def get_extensions():
     -------
     List[`asdf.extension.Extension`]
     """
-    from ._converters import NODE_EXTENSIONS
+    from .converters import NODE_EXTENSIONS
 
     return NODE_EXTENSIONS
