+++ src/roman_datamodels/datamodels/_core.py
@@ -21,6 +21,7 @@ from asdf.exceptions import ValidationError
 from astropy.time import Time
 
 from roman_datamodels import stnode, validate
+from roman_datamodels._asdf.converters import NODE_EXTENSIONS
 
 __all__ = ["DataModel", "MODEL_REGISTRY"]
 
@@ -162,7 +163,7 @@ class DataModel(abc.ABC):
     @property
     def schema_uri(self):
         # Determine the schema corresponding to this model's tag
-        return next(t for t in stnode.NODE_EXTENSIONS[0].tags if t.tag_uri == self._instance._tag).schema_uris[0]
+        return next(t for t in NODE_EXTENSIONS[0].tags if t.tag_uri == self._instance._tag).schema_uris[0]
 
     def close(self):
         if not (self._iscopy or self._asdf is None):
