+++ src/roman_datamodels/stnode/__init__.py
@@ -1,8 +1,3 @@
-"""
-The STNode classes and supporting objects generated dynamically at import time
-    from RAD's manifest.
-"""
-from ._converters import *  # noqa: F403
 from ._node import *  # noqa: F403
 from ._nodes import *  # noqa: F403
 from ._tagged import *  # noqa: F403
