+++ tests/test_models.py
@@ -12,6 +12,7 @@ from numpy.testing import assert_array_equal
 from roman_datamodels import datamodels
 from roman_datamodels import maker_utils as utils
 from roman_datamodels import stnode, validate
+from roman_datamodels._asdf.converters import NODE_EXTENSIONS
 from roman_datamodels.testing import assert_node_equal
 
 from .conftest import MANIFESTS
@@ -327,7 +328,7 @@ def test_reference_file_model_base(tmp_path):
     # Set temporary asdf file
 
     # Get all reference file classes
-    tags = [t for t in stnode.NODE_EXTENSIONS[0].tags if "/reference_files/" in t.tag_uri]
+    tags = [t for t in NODE_EXTENSIONS[0].tags if "/reference_files/" in t.tag_uri]
     for tag in tags:
         schema = asdf.schema.load_schema(tag.schema_uris[0])
         # Check that schema references common reference schema
